lib/igt_sriov_device: Adjust the igt_sriov_is_pf function
A device is considered a PF if it supports the creation of one or more
VFs. This capability is indicated by the presence of the sriov_totalvfs
attribute in the device's sysfs entry, with a value greater than zero.
The igt_sriov_is_pf function checks for this attribute and its value to
confirm the device's PF status.
This adjustment streamlines the process for tests that manipulate Virtual
Functions (VFs) by consolidating prerequisite checks into a single
`igt_require(igt_sriov_is_pf)`.

Cc: Lukasz Laguna <lukasz.laguna@intel.com>
Cc: Michal Wajdeczko <michal.wajdeczko@intel.com>
Signed-off-by: Marcin Bernatowicz <marcin.bernatowicz@linux.intel.com>
Reviewed-by: Adam Miszczak <adam.miszczak@linux.intel.com>