lib/igt_psr: Adjust igt_wait intervals
Currently we are observing problems with psr_wait_update. Poll interval is
10 ms. With eDP Panel Replay we are seeing scenario where on first poll PSR
status haven't yet changed to !SLEEP. Then we are waiting 10ms before next
check. During this 10ms sleep status changes to !SLEEP and back to SLEEP
and this is not noticed by psr_wait update and causing failure.

Fix this by adjusting poll intervals to 1ms.

Signed-off-by: Jouni Högander <jouni.hogander@intel.com>
Reviewed-by: Kunal Joshi <kunal1.joshi@intel.com>