lib/rendercopy: Add specific support for 2:10:10:10 formats
Use B10G10R10A2_UNORM instead of B8G8R8A8_UNORM when dealing
with actual 10bpc pixel formats.

This is needed on tgl+ because the display hardware decompressor
expects some magic bit shuffling to have taken place. If the
compressor didn't do that we get garbage.

Also if the clear color is involved then the hardware needs
to know the actual pixel format in order to correctly generate
the native version of the clear color (which will be consumed
by the display hardware.

Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>