lib/igt_sriov_device: Replace libpciaccess with direct sysfs interaction
Previously, libpciaccess was used to fetch PCI device information for
VFs in SR-IOV, which does not refresh the PCI bus state after the initial
scan. This can lead to outdated PCI information in dynamic SR-IOV
environments where VFs are dynamically managed.

The new implementation directly interacts with sysfs to resolve the PCI
slot address of a VF, ensuring access to the most current state of the
PCI bus. This change removes the dependency on libpciaccess and uses
__igt_sriov_get_vf_pci_slot_alloc to fetch the PCI slot addresses
directly from sysfs based on VF numbers.

Reported-by: Michal Wajdeczko <michal.wajdeczko@intel.com>
Closes: https://intel-gfx-ci.01.org/tree/intel-xe/xe-pw-135423v1/shard-adlp-9/igt@sriov_basic@bind-unbind-vf@vf-2.html
Cc: Michal Wajdeczko <michal.wajdeczko@intel.com>
Cc: Lukasz Laguna <lukasz.laguna@intel.com>
Cc: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Signed-off-by: Marcin Bernatowicz <marcin.bernatowicz@linux.intel.com>
Reviewed-by: Lukasz Laguna <lukasz.laguna@intel.com>