lib/xe/oa: Add OA LNL metrics (oa_lnl.xml)
$ mdapi-xml-convert.py --guids=guids.xml MetricsXML_P31.xml > oa-lnl.xml

Note: The metrics XML's consist of (a) the registers (in this case NOA or
mux registers) to be programmed (b) the metrics themselves which specify
how the counters in the reports should be combined to construct the
metrics.  The changes in mdapi-xml-convert.py in this patch completely
ignore (b) since they are not relevant for IGT. (a) is needed for IGT LNL
OAG uses "576B_PEC64LL" format which is not available on previous platforms
so we cannot use metrics for a previous platform such as MTL to generate
LNL OAG data.

Note: LNL media guids had issues after running mdapi-convert-xml.py so have
not been added, revisit this.

Signed-off-by: Ashutosh Dixit <ashutosh.dixit@intel.com>
Reviewed-by: Umesh Nerlige Ramappa <umesh.nerlige.ramappa@intel.com>