tests/intel/xe_evict: Reduce the "large" bo size for threaded eviction
When calculating the number of bos that simultaneously fits in
VRAM + system memory, account for eviction pipelining by subtracting
one bo for ongoing pipelined evictions per thread.

With large bos this may lead to us not being able to use a working
set of 2 bos, which is the minimum, so reduce the large bo size for
threaded evictions, and instead increase the default number of bos.

v3:
- Fix up number of bo calculation to *really* subtract one bo per
  thread. (Matthew Brost)

Signed-off-by: Thomas Hellström <thomas.hellstrom@linux.intel.com>
Reviewed-by: Nirmoy Das <nirmoy.das@intel.com>