lib/vebox: Document that 2:10:10:10 is unusable with VEBOX
VEBOX has a specific format for R10G10B10A2_UNORM, but it
just hangs the engine if you try to use it.

For most platforms bsepc says: "Valid format only for when
IECP is enabled and in output state only. Not supported in
HSB mode."

Document this fact so that the next fella doesn't try to
implement it like I did :/

Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>