lib/igt_fb: Assert that we have no clear color when using the bltter
The blitter code doesn't know how to deal with the clear color
currently. There is some preliminary code to simply skip the clear
color buffer for dg2, but that would just result in bogus rendering
if we even had to deal with fast cleared stuff.

Rip out the dg2 stuff, and just assert that we should never get
here with a clear color modifier.

Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>