lib/igt_fb: Try to fix block copy media compression handling
The media compression format handling for the block
copy seems to be very broken.

Empirical findings from dg2:
- the compression format needs to be based on bspec's
  render compression format list even when doing media
  compression
- 0x8/B8G8G8A8 doesn't work correctly for XYUV8888 for
  some reason, but using 0x18/R8 does seem to work
- YUYV/etc. packed formats don't work any which way
- planar formats can be handled as R8/R8G8/etc.
- RGB+media compression is a lost cause (black turns white
  when you try it)

This is all based on very cursory testing, so no idea if any
of it is actually correct. Doesn't actually matter currently
as we take the enginecopy path for media compression always.

Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>