lib/xe/oa: Handle SIMD16_EU_PER_DSS
Handle SIMD16_EU_PER_DSS the same way EU_PER_DSS is. Since this function
is mixing information from hwconfig and the topo query, this should be
more correct as now the numbers from the query match the hwconfig.

This fixes the assert triggering in the default case because
SIMD16_EU_PER_DSS was not handled:

	(xe_oa:4511) CRITICAL: Test assertion failure function xe_fill_topology_info, file ../lib/xe/xe_oa.c:532:
	(xe_oa:4511) CRITICAL: Failed assertion: 0
	(xe_oa:4511) CRITICAL: Last errno: 2, No such file or directory
	Test xe_oa failed.

Cc: Ashutosh Dixit <ashutosh.dixit@intel.com>
Reviewed-by: Ashutosh Dixit <ashutosh.dixit@intel.com>
Link: https://lore.kernel.org/r/20240722190717.57192-1-lucas.demarchi@intel.com
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>