tests/intel/xe_exec_compute_mode: Add userptr free / munmap sections
Add userptr free / munmap sections which does the aforementioned while
the userptr still has a mapping. The userptr will be invalidated
triggering a rebind which will fail pinning pages as the userptr is no
longer valid. This is allowed and should not corrupt the VM or future
submissions.

v2:
 - s/faill/fail (Janga)
 - check dummy for mapped failed (Janga)
v3:
 - Include local changes

Cc: Janga Rahul Kumar <janga.rahul.kumar@intel.com>
Signed-off-by: Matthew Brost <matthew.brost@intel.com>
Reviewed-by: Janga Rahul Kumar <janga.rahul.kumar@intel.com>