benchmarks/gem_wsim: Introduce intel_engines structure
Introduction of struct intel_engines, which encapsulates the number of
engines (nr_engines) and a pointer to an array of engine IDs (engines).
This structural refactor replaces the previous ad-hoc approach of managing
engine maps across various structures (w_step, ctx, etc.)
This change is part of a series of preparatory steps for upcoming
patches.

v2: Correct indentation.

Signed-off-by: Marcin Bernatowicz <marcin.bernatowicz@linux.intel.com>
Acked-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>