benchmarks/gem_wsim: Update request_idx in prepare phase
Renamed `request` to `request_idx` for consistency with `engine_idx`.
The index references an element of device's physical engines array.
Similar to engine_idx the field is initialized in workload's preparation
phase. Field is used for throttling functionality, enabling control over
the rate of requests on a given engine.

v2: Avoid multiple assignments.

Signed-off-by: Marcin Bernatowicz <marcin.bernatowicz@linux.intel.com>
Acked-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>