tests/intel-ci/xe: Block kms_rmfb@close-fd test
Upon reopening the driver, the 'kms_rmfb@close-fd' test expects
that 'getcrtc' should return an FB ID of 0. However, due to the
default modeset, 'getcrtc' always returns a non-zero FB ID, which
is expected behavior.

In the case of the i915 driver, the 'i915_gem_drop_caches' clears
the cache and sets the FB ID to 0. Currently, there is no equivalent
mechanism in the XE driver.

Therefore, we should block this test from XE execution until an
alternative solution is implemented.

Closes: https://gitlab.freedesktop.org/drm/xe/kernel/-/issues/294
Signed-off-by: Bhanuprakash Modem <bhanuprakash.modem@intel.com>
Reviewed-by: Jeevan B <jeevan.b@intel.com>