tests/amdgpu: enable amd_deadlock test on Renior
When family_id is AMDGPU_FAMILY_RV, we further check if it is Renoir
by chip_external_rev.

When checking if a command ring is available on a specific AMD GPU in
is_rings_available(), delete the igt_assert_eq(). Because it is normal
that not all the rings are available on all ASIC families.
For example, there is no AMD_IP_VPE ring on Renior platform

V2: Using memset() instead of zero initialization to initialize
drm_amdgpu_info_hw_ip to avoid compiler compliance issues(Vitaly)

Signed-off-by: Trigger Huang <Trigger.Huang@amd.com>
Reviewed-by: Vitaly Prosyak <vitaly.prosyak@amd.com>