lib/xe/xe_query: Do not require QUERY_OA_UNITS for all Xe tests
The code currently requires the driver to support DRM_XE_DEVICE_QUERY_OA_UNITS
to run all tests while the only test that needs it is xe_oa. Removes the
requirement so that all other Xe tests can run when this single requirement is
not met.

I found it when trying to run IGT Xe tests on a recent kernel
6.9.9-200.fc40.x86_64 from Fedora 40.

V4:
 - Shorten the commit message
 - Adds back the the path prefix to the subject

V3:
 - Update comment syntax and shorten comments
 - Update comment content to describe from kernel version perspective only
 - Update test to make '!= 0' implicit

V2:
 - Cleanup headers from message body
 - Add 'lib/xe/xe_query:' to subject line

Reviewed-by: Ashutosh Dixit <ashutosh.dixit@intel.com>
Signed-off-by: Peter Senna Tschudin <peter.senna@intel.com>