tests/intel/xe_query: Fail on unknown hwconfig entries
There is a test that the hwconfig table is valid. However, if there
was an unknown entry (corrupted table or out of date test) all it did
was print a debug message of "please fix"! So the test would still
pass and CI would not notice the problem.

Add in some actual asserts that the enum value is within the valid
range and that it is not missing from the switch statement.

Signed-off-by: John Harrison <John.C.Harrison@Intel.com>
Reviewed-by: Ashutosh Dixit <ashutosh.dixit@intel.com>