intel-gfx-fw-info: Allow to print checksum
Choose a checksum algo and use it to print when parsing a firmware.
This may be useful in CI to dump what is the exact firmware used
and may also be useful when inspecting the firmware repo and making
sure mistakes are not introduced. Configuring intel-gfx-fw-info a diff
driver:

	firmware $ echo "*.bin   diff=intelfw" >> .git/info/attributes
	firmware $ git config --local \
		diff.intelfw.textconv \
		"<path-to-igt>/tools/intel-gfx-fw-info -c"

Example for the LNL update that happened recently:

	firmware $ git diff --text bc42cdc34bd8 b5bef5bdbab1 --  xe/lnl_guc_70.bin
	diff --git a/xe/lnl_guc_70.bin b/xe/lnl_guc_70.bin
	index 785763163..75a1958f7 100644
	--- a/xe/lnl_guc_70.bin
	+++ b/xe/lnl_guc_70.bin
	@@ -1,3 +1,3 @@
	-version: 70.20.0
	-date: 2024-02-09
	-checksum: 0809f459048a23716fd3019074f36812163e027835cbb43acfcfe72cd83fb0cc
	+version: 70.29.2
	+date: 2024-07-26
	+checksum: e7ceae28b06bd71c53fe177f4f6787aac88956d8944fa02b1ab2ec53b1e961e4

If for some reason there's only a checksum change, something went wrong,
and without it we wouldn't notice when using this tool as a diff driver.

Reviewed-by: Gustavo Sousa <gustavo.sousa@intel.com>
Link: https://lore.kernel.org/r/20240813152141.3151865-2-lucas.demarchi@intel.com
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>