lib/xe: Remove MS_TO_NS()
This was added in lib/xe/xe_query.h to quickly convert the kernel
interface to use nsec rather than msec. However it's misplaced and
doesn't follow the same style as for other time-conversion macros.

Worth noting that the "#define ONE_SEC" and the like are error-prone
if we are using other interfaces with different units. Example:
`usleep(ONE_SEC)` could go unnoticed since there's no unit in the
define.

Use the common way of doing time-conversion via the NSEC_PER_SEC
macros.

Reviewed-by: Gustavo Sousa <gustavo.sousa@intel.com>
Link: https://lore.kernel.org/r/20240814214705.3718355-3-lucas.demarchi@intel.com
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>