tools/intel-gfx-fw-info: Embed fw decoding
Use a factory class method to decide what class to use depending on the
magic and embed the calls to cstruct in each of them. Main motivation is
that in future the decode method will need to seek the binary as it
can't really know the version just checking the beginning of the file.

For now, keep the open/close of the file in the main function, but
eventually can be migrated as well.

v2 (Gustavo):
- Fix s/kw/**kw/
- Remove unnecessary seek()
- Use a constructor variable in create()
- Squash call to sys.exit() that was in the wrong patch
v3 (Gustavo):
- Fix f-string

Acked-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Reviewed-by: Gustavo Sousa <gustavo.sousa@intel.com>
Link: https://lore.kernel.org/r/20240826152656.1306490-4-lucas.demarchi@intel.com
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>