lib/igt_fb: fix stride calculation for igt_fb_get_fnv1a_crc()
fb->strides is assigned to the value returned by
calc_plane_stride() only if it was previously not set by the
caller. In the case of dumb buffer allocation, this is already
assigned by the driver.

Respect the stride requested by the driver in such cases by
using fb->strides[0] as this is assigned correctly in
calc_fb_size() only if it was previously not set by the driver.

Signed-off-by: Esha Bharadwaj <quic_ebharadw@quicinc.com>