gputop: Print message without clients
Devs are often confused why gputop doesn't display anything and think it
doesn't yet support the driver/platform. This is different from
intel_gpu_top that can display a header with engines, because in that
case there is a fixed set of engines and no variance according to the
driver.

Print a message to give a hint that nothing is displayed because there's
nothing running.

Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Link: https://lore.kernel.org/r/20240912141426.4115404-1-lucas.demarchi@intel.com
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>