tests/kms_display_modes: Fix the plane setup in extended-mode-basic
The test forgets to set the output size of the planes. The
default plane output size will be taken from the fb dimensions
which means we end up upscaling the planes, which may not work
on all hardware.

Set the planes' output size to match the input size to avoid this.
The frambuffer was specifically created wide enough to cover both
outputs with 1:1 scaling.

Closes: https://gitlab.freedesktop.org/drm/intel/-/issues/9841
Reviewed-by: Bhanuprakash Modem <bhanuprakash.modem@intel.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Signed-off-by: Bhanuprakash Modem <bhanuprakash.modem@intel.com> (Re-float)