tests: drm_fdinfo: Fix zero tolerance checks
When we expect an engine to be busy, we check if its reported busy time
falls within a +/-5% tolerance range of measurement time period.
However, when we expect the engine to be idle, we compare its reported
busy time against zero, still with a +/-5% tolerance range, but now
calculated against the zero value, then no tolerance at all.  Obviously,
such check fails when the reported busy time is not exactly zero.

Compare engine idle time against measurement time period instead of
comparing its busy time against zero when we expect the busy time to be
close to zero.  As a debugging aid, display messages with the compared
values when requested via --debug option or when a failure occurs.

Closes: https://gitlab.freedesktop.org/drm/i915/kernel/-/issues/7742
Suggested-by: Chris Wilson <chris.p.wilson@linux.intel.com>
Reviewed-by: Peter Senna Tschudin <peter.senna@linux.intel.com>
Signed-off-by: Janusz Krzysztofik <janusz.krzysztofik@linux.intel.com>