lib/igt_kmod: Rewrite xe unload logic
Stop trying to unload possibly-dependent modules. Loading a module and
probing a module are 2 different things, although the Linux kernel will
by default probe devices when a module is loaded (for pci, configurable
in /sys/bus/pci/drivers_autoprobe).

For the unloading part we can simplify: instead of trying to unload
the modules directly (which may complain that module is already in use)
we just unplug the HW first.  For each possible device bound.

This will become part of libkmod in future, but we don't need to wait:
we can implement it in igt too as it's straightforward.

For now this only switches the unload path for xe, but if it works well,
we can switch others in future too.

v2:
- Rename function since now it works for any PCI driver
- Move function so it will be available for i915 too.

Closes: https://gitlab.freedesktop.org/drm/xe/kernel/-/issues/2362
Reviewed-by: Jonathan Cavitt <jonathan.cavitt@intel.com>
Link: https://lore.kernel.org/r/20240919151130.1410357-1-lucas.demarchi@intel.com
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>