tests/kms_rmfb: Fix support for xe
Make sure to hold a drm fd open while running the test so we can check
for fb_id == 0 after either removing the fb or closing the previous fd.
With i915 this is guaranteed by holding on to an fd to be used in an
exit handler in drm_open_driver().

Cc: Bhanuprakash Modem <bhanuprakash.modem@intel.com>
Cc: Swati Sharma <swati2.sharma@intel.com>
Reviewed-by: Jeevan B <jeevan.b@intel.com>
Link: https://lore.kernel.org/r/20240919193250.1439745-2-lucas.demarchi@intel.com
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>