lib/xe_eudebug: Increase default pipe timeout value
Our discovery tests are quite heavy, there can be tens of
threads of clients and debuggers racing against eachothers.

If there are lots of debugs enabled in kernel config, it is
in real of possibility to end up in a situation where one
client or one debugger really was starved more than 25 seconds.

This timeout is used in communication between test framework,
debuggers and debuggees and thus if we timeout on it,
it will lead to confusing errors in middle tests.

The tradeoff is that for really failing tests where the
other endpoint ceases to respond, we will detect that now
slower.

Signed-off-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>
Reviewed-by: Dominik Grzegorzek <dominik.grzegorzek@intel.com>